requirements.txt
# AI Safety Copilot Requirements

# Core AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.10.0
anthropic>=0.18.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pypdf>=3.17.0

# Web Framework
streamlit>=1.29.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.1.0

# Logging
python-json-logger>=2.0.0

# Optional: Local LLM support
# ollama>=0.1.7  # Uncomment if using local Llama 3

